oh hey analyzerlib typechecks now
