+++ src/python/grapl_analyzerlib/grapl_analyzerlib/nodes/engagement.py
@@ -1,4 +1,6 @@
+from __future__ import annotations
 import logging
+import os
 import sys
 
 from typing import Any, Dict, List, Tuple, Union, Optional, cast, TypeVar, Set
