+++ src/python/grapl_analyzerlib/grapl_analyzerlib/nodes/lens.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import json
 
 from typing import Any, TypeVar, List, Set, Dict, Tuple, Optional
