+++ src/python/grapl_analyzerlib/grapl_analyzerlib/nodes/network_connection.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from typing import Any, TypeVar, Set, Dict, Tuple, Optional
 
 from grapl_analyzerlib.node_types import (
@@ -9,6 +10,7 @@ from grapl_analyzerlib.node_types import (
 from grapl_analyzerlib.nodes.entity import EntityQuery, EntityView, EntitySchema
 from grapl_analyzerlib.queryable import with_str_prop, with_int_prop
 from grapl_analyzerlib.schema import Schema
+from grapl_analyzerlib.comparators import IntOrNot, StrOrNot, OneOrMany
 
 NCQ = TypeVar("NCQ", bound="NetworkConnectionQuery")
 NCV = TypeVar("NCV", bound="NetworkConnectionView")
@@ -55,11 +57,11 @@ class NetworkConnectionQuery(EntityQuery[NCV, NCQ]):
     def with_port(
         self,
         *,
-        eq: Optional["IntOrNot"] = None,
-        gt: Optional["IntOrNot"] = None,
-        ge: Optional["IntOrNot"] = None,
-        lt: Optional["IntOrNot"] = None,
-        le: Optional["IntOrNot"] = None,
+        eq: Optional[IntOrNot] = None,
+        gt: Optional[IntOrNot] = None,
+        ge: Optional[IntOrNot] = None,
+        lt: Optional[IntOrNot] = None,
+        le: Optional[IntOrNot] = None,
     ):
         pass
 
@@ -67,11 +69,11 @@ class NetworkConnectionQuery(EntityQuery[NCV, NCQ]):
     def with_ip_address(
         self,
         *,
-        eq: Optional["StrOrNot"] = None,
-        contains: Optional["OneOrMany[StrOrNot]"] = None,
-        starts_with: Optional["StrOrNot"] = None,
-        ends_with: Optional["StrOrNot"] = None,
-        regexp: Optional["OneOrMany[StrOrNot]"] = None,
+        eq: Optional[StrOrNot] = None,
+        contains: Optional[OneOrMany[StrOrNot]] = None,
+        starts_with: Optional[StrOrNot] = None,
+        ends_with: Optional[StrOrNot] = None,
+        regexp: Optional[OneOrMany[StrOrNot]] = None,
         distance_lt: Optional[Tuple[str, int]] = None,
     ):
         pass
