+++ src/python/grapl_analyzerlib/grapl_analyzerlib/retry.py
@@ -1,6 +1,8 @@
 import time
 from functools import wraps
-from typing import Type
+from typing import Callable, Type, TypeVar, cast
+
+F = TypeVar("F", bound=Callable)
 
 
 def retry(
@@ -9,7 +11,7 @@ def retry(
     tries: int = 3,
     delay: float = 0.5,
     backoff: int = 2,
-):
+) -> Callable[[F], F]:
     """Retry calling the decorated function using an exponential backoff.
 
     Modified to handle 'falsey' cases with a retry
@@ -30,7 +32,7 @@ def retry(
         each retry
     """
 
-    def deco_retry(f):
+    def deco_retry(f: F) -> F:
         @wraps(f)
         def f_retry(*args, **kwargs):
             mtries, mdelay = tries, delay
